+++ test/unittest/inner_api/src/user_auth_callback_proxy_test.cpp
@@ -49,7 +49,7 @@ HWTEST_F(UserAuthCallbackProxyTest, TestOnResult_001, TestSize.Level0)
     EXPECT_CALL(*obj, SendRequest(_, _, _, _))
         .WillOnce(
             [](uint32_t code, MessageParcel &data, MessageParcel &reply, MessageOption &option) {
-                EXPECT_EQ(code, static_cast<uint32_t>(UserAuthInterfaceCode::USER_AUTH_ON_RESULT));
+                EXPECT_EQ(code, UserAuthInterfaceCode::USER_AUTH_ON_RESULT);
                 return OHOS::NO_ERROR;
             }
         );
@@ -69,7 +69,7 @@ HWTEST_F(UserAuthCallbackProxyTest, TestOnAcquireInfo_001, TestSize.Level0)
     EXPECT_CALL(*obj, SendRequest(_, _, _, _))
         .WillOnce(
             [](uint32_t code, MessageParcel &data, MessageParcel &reply, MessageOption &option) {
-                EXPECT_EQ(code, static_cast<uint32_t>(UserAuthInterfaceCode::USER_AUTH_ACQUIRE_INFO));
+                EXPECT_EQ(code, UserAuthInterfaceCode::USER_AUTH_ACQUIRE_INFO);
                 return OHOS::NO_ERROR;
             }
         );
@@ -90,7 +90,7 @@ HWTEST_F(UserAuthCallbackProxyTest, TestOnGetExecutorPropertyResult_001, TestSiz
     EXPECT_CALL(*obj, SendRequest(_, _, _, _))
         .WillOnce(
             [](uint32_t code, MessageParcel &data, MessageParcel &reply, MessageOption &option) {
-                EXPECT_EQ(code, static_cast<uint32_t>(UserAuthInterfaceCode::USER_AUTH_GET_EX_PROP));
+                EXPECT_EQ(code, UserAuthInterfaceCode::USER_AUTH_GET_EX_PROP);
                 return OHOS::NO_ERROR;
             }
         );
